Tidying methods for ETS (Error, Trend, Seasonal) exponential smoothing modeling of time series

Tidying methods for ETS (Error, Trend, Seasonal) exponential smoothing modeling of time series

# S3 method for ets
sw_tidy(x, ...)

# S3 method for ets
sw_glance(x, ...)

# S3 method for ets
sw_augment(x, data = NULL, timetk_idx = FALSE,
  rename_index = "index", ...)

# S3 method for ets
sw_tidy_decomp(x, timetk_idx = FALSE, rename_index = "index",
  ...)

Arguments

x
An object of class "ets"

...
Not used.

data
Used with sw_augment only. NULL by default which simply returns augmented columns only. User can supply the original data, which returns the data + augmented columns.

timetk_idx
Used with sw_augment and sw_tidy_decomp. When TRUE, uses a timetk index (irregular, typically date or datetime) if present.

rename_index
Used with sw_augment and sw_tidy_decomp. A string representing the name of the index generated.

Value

sw_tidy() returns one row for each model parameter, with two columns:

term: The smoothing parameters (alpha, gamma) and the initial states (l, s0 through s10)

estimate: The estimated parameter value

sw_glance() returns one row with the columns

model.desc: A description of the model including the three integer components (p, d, q) are the AR order, the degree of differencing, and the MA order.

sigma: The square root of the estimated residual variance

logLik: The data's log-likelihood under the model

AIC: The Akaike Information Criterion

BIC: The Bayesian Information Criterion

ME: Mean error

RMSE: Root mean squared error

MAE: Mean absolute error

MPE: Mean percentage error

MAPE: Mean absolute percentage error

MASE: Mean absolute scaled error

ACF1: Autocorrelation of errors at lag 1

sw_augment() returns a tibble with the following time series attributes:

index: An index is either attempted to be extracted from the model or a sequential index is created for plotting purposes

.actual: The original time series

.fitted: The fitted values from the model

.resid: The residual values from the model

sw_tidy_decomp() returns a tibble with the following time series attributes:

index: An index is either attempted to be extracted from the model or a sequential index is created for plotting purposes

observed: The original time series

level: The level component

slope: The slope component (Not always present)

season: The seasonal component (Not always present)

See also

ets()

Examples

library(forecast)
library(sweep)

fit_ets <- WWWusage %>%
    ets()

sw_tidy(fit_ets)

#> # A tibble: 5 x 2
#>    term   estimate
#>   <chr>      <dbl>
#> 1 alpha  0.9999000
#> 2  beta  0.9999000
#> 3   phi  0.8028209
#> 4     l 84.6792445
#> 5     b  1.7805641

sw_glance(fit_ets)

#> # A tibble: 1 x 12
#>    model.desc    sigma    logLik      AIC      BIC        ME     RMSE      MAE
#>         <chr>    <dbl>     <dbl>    <dbl>    <dbl>     <dbl>    <dbl>    <dbl>
#> 1 ETS(A,Ad,N) 3.461057 -354.4159 720.8318 736.4629 0.2364937 3.461057 2.798415
#> # ... with 4 more variables: MPE <dbl>, MAPE <dbl>, MASE <dbl>, ACF1 <dbl>

sw_augment(fit_ets)

#> # A tibble: 100 x 4
#>    index .actual  .fitted     .resid
#>    <int>   <dbl>    <dbl>      <dbl>
#>  1     1      88 86.10872  1.8912815
#>  2     2      84 90.66563 -6.6656307
#>  3     3      85 80.79007  4.2099294
#>  4     4      85 85.80153 -0.8015264
#>  5     5      84 85.00048 -1.0004826
#>  6     6      85 83.19730  1.8027048
#>  7     7      83 85.80242 -2.8024154
#>  8     8      85 81.39501  3.6049916
#>  9     9      88 86.60477  1.3952335
#> 10    10      89 90.40850 -1.4085004
#> # ... with 90 more rows

sw_tidy_decomp(fit_ets)

#> # A tibble: 101 x 4
#>    index observed    level         slope
#>    <dbl>    <dbl>    <dbl>         <dbl>
#>  1     0       NA 84.67924  1.7805640880
#>  2     1       88 87.99981  3.3205662984
#>  3     2       84 84.00067 -3.9991440097
#>  4     3       85 84.99958  0.9989120994
#>  5     4       85 85.00008  0.0005012949
#>  6     5       84 84.00010 -0.9999800904
#>  7     6       85 84.99982  0.9997195865
#>  8     7       83 83.00028 -1.9995393336
#>  9     8       85 84.99964  1.9993590469
#> 10     9       88 87.99986  3.0002210271
#> # ... with 91 more rows