Tidying methods for STL (Seasonal, Trend, Level) decomposition of time series

Tidying methods for STL (Seasonal, Trend, Level) decomposition of time series

# S3 method for stl
sw_tidy(x, ...)

# S3 method for stl
sw_tidy_decomp(x, timetk_idx = FALSE, rename_index = "index",
  ...)

# S3 method for stlm
sw_tidy_decomp(x, timetk_idx = FALSE, rename_index = "index",
  ...)

Arguments

x
An object of class "stl" or "stlm"

...
Not used.

timetk_idx
Used with sw_tidy_decomp. When TRUE, uses a timetk index (irregular, typically date or datetime) if present.

rename_index
Used with sw_tidy_decomp. A string representing the name of the index generated.

Value

sw_tidy() wraps sw_tidy_decomp() sw_tidy_decomp() returns a tibble with the following time series attributes:

index: An index is either attempted to be extracted from the model or a sequential index is created for plotting purposes

season: The seasonal component

trend: The trend component

remainder: observed - (season + trend)

seasadj: observed - season (or trend + remainder)

See also

stl()

Examples

library(forecast)
library(sweep)

fit_stl <- USAccDeaths %>%
    stl(s.window = "periodic")

sw_tidy_decomp(fit_stl)

#> # A tibble: 72 x 6
#>            index observed      season    trend  remainder   seasadj
#>    <S3: yearmon>    <dbl>       <dbl>    <dbl>      <dbl>     <dbl>
#>  1      Jan 1973     9007  -819.87198 9934.537 -107.66547  9826.872
#>  2      Feb 1973     8106 -1559.05452 9880.772 -215.71718  9665.055
#>  3      Mar 1973     8928  -759.57068 9827.006 -139.43526  9687.571
#>  4      Apr 1973     9137  -530.48523 9765.695  -98.21000  9667.485
#>  5      May 1973    10017   334.60040 9704.385  -21.98493  9682.400
#>  6      Jun 1973    10826   814.91223 9636.872  374.21567 10011.088
#>  7      Jul 1973    11317  1681.72424 9569.360   65.91609  9635.276
#>  8      Aug 1973    10744   982.03921 9500.044  261.91646  9761.961
#>  9      Sep 1973     9713   -62.81261 9430.729  345.08361  9775.813
#> 10      Oct 1973     9938   231.80889 9343.281  362.90988  9706.191
#> # ... with 62 more rows