Validates data frame has column named the same name as variable rename_index

Validates data frame has column named the same name as variable rename_index

Usage

validate_index(ret, rename_index)

Arguments

ret

An object of class tibble

rename_index

A variable indicating the index name to be used in the tibble returned