Coerces decomposed time-series objects to tibble format.

Coerces decomposed time-series objects to tibble format.

sw_tidy_decomp(x, timetk_idx = FALSE, rename_index = "index", ...)

Arguments

x
A time-series object of class stl, ets, decomposed.ts, HoltWinters, bats or tbats.

timetk_idx
When TRUE, uses a timetk index (irregular, typically date or datetime) if present.

rename_index
Enables the index column to be renamed.

...
Not used.

Value

Returns a tibble object.

Details

sw_tidy_decomp is designed to coerce time-series objects with decompositions to tibble objects.

A regularized time index is always constructed. If no time index is detected, a sequential index is returned as a default. The index column name can be changed using the rename_index argument.

Examples

library(forecast)
library(sweep)

# Decompose ETS model
USAccDeaths %>%
    ets() %>%
    sw_tidy_decomp()

#> # A tibble: 73 x 4
#>            index observed    level      season
#>    <S3: yearmon>    <dbl>    <dbl>       <dbl>
#>  1      Dec 1972       NA 9196.970   -50.32965
#>  2      Jan 1973     9007 9561.681  -809.05699
#>  3      Feb 1973     8106 9602.363 -1524.73692
#>  4      Mar 1973     8928 9662.423  -776.31385
#>  5      Apr 1973     9137 9650.510  -505.20082
#>  6      May 1973    10017 9660.347   349.79263
#>  7      Jun 1973    10826 9903.851   752.31183
#>  8      Jul 1973    11317 9758.817  1659.33955
#>  9      Aug 1973    10744 9753.485   994.23405
#> 10      Sep 1973     9713 9781.133   -87.41629
#> # ... with 63 more rows

# Decompose STL object
USAccDeaths %>%
    stl(s.window = 'periodic') %>%
    sw_tidy_decomp()

#> # A tibble: 72 x 6
#>            index observed      season    trend  remainder   seasadj
#>    <S3: yearmon>    <dbl>       <dbl>    <dbl>      <dbl>     <dbl>
#>  1      Jan 1973     9007  -819.87198 9934.537 -107.66547  9826.872
#>  2      Feb 1973     8106 -1559.05452 9880.772 -215.71718  9665.055
#>  3      Mar 1973     8928  -759.57068 9827.006 -139.43526  9687.571
#>  4      Apr 1973     9137  -530.48523 9765.695  -98.21000  9667.485
#>  5      May 1973    10017   334.60040 9704.385  -21.98493  9682.400
#>  6      Jun 1973    10826   814.91223 9636.872  374.21567 10011.088
#>  7      Jul 1973    11317  1681.72424 9569.360   65.91609  9635.276
#>  8      Aug 1973    10744   982.03921 9500.044  261.91646  9761.961
#>  9      Sep 1973     9713   -62.81261 9430.729  345.08361  9775.813
#> 10      Oct 1973     9938   231.80889 9343.281  362.90988  9706.191
#> # ... with 62 more rows