Tidying methods for StructTS (Error, Trend, Seasonal) / exponential smoothing modeling of time series

These methods tidy the coefficients of StructTS models of univariate time series.

# S3 method for StructTS
sw_tidy(x, ...)

# S3 method for StructTS
sw_glance(x, ...)

# S3 method for StructTS
sw_augment(x, data = NULL, timetk_idx = FALSE, rename_index = "index", ...)

Arguments

x
An object of class "StructTS"

...
Additional parameters (not used)

data
Used with sw_augment only. NULL by default which simply returns augmented columns only. User can supply the original data, which returns the data + augmented columns.

timetk_idx
Used with sw_augment only. Uses a irregular timetk index if present.

rename_index
Used with sw_augment only. A string representing the name of the index generated.

Value

sw_tidy() returns one row for each model parameter, with two columns:

term: The model parameters

estimate: The estimated parameter value

sw_glance() returns one row with the columns

model.desc: A description of the model including the three integer components (p, d, q) are the AR order, the degree of differencing, and the MA order.

sigma: The square root of the estimated residual variance

logLik: The data's log-likelihood under the model

AIC: The Akaike Information Criterion

BIC: The Bayesian Information Criterion

ME: Mean error

RMSE: Root mean squared error

MAE: Mean absolute error

MPE: Mean percentage error

MAPE: Mean absolute percentage error

MASE: Mean absolute scaled error

ACF1: Autocorrelation of errors at lag 1

sw_augment() returns a tibble with the following time series attributes:

index: An index is either attempted to be extracted from the model or a sequential index is created for plotting purposes

.actual: The original time series

.fitted: The fitted values from the model

.resid: The residual values from the model

See also

Examples

library(dplyr)
library(forecast)
library(sweep)

fit_StructTS <- WWWusage %>%
    StructTS()

sw_tidy(fit_StructTS)

#> # A tibble: 3 x 2
#>   term    estimate
#>   <chr>      <dbl>
#> 1 level        0  
#> 2 slope       13.0
#> 3 epsilon      0  

sw_glance(fit_StructTS)

#> # A tibble: 1 x 12
#>   model.desc sigma logLik   AIC   BIC       ME  RMSE   MAE    MPE  MAPE  MASE
#>   <chr>      <dbl>  <dbl> <dbl> <dbl>    <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl>
#> 1 Local lin… 0.995  -277.  559.  564. -0.00554 0.995 0.820 0.0389 0.641 0.181
#> # … with 1 more variable: ACF1 <dbl>

sw_augment(fit_StructTS)

#> Warning: `select_()` is deprecated as of dplyr 0.7.0.
#> Please use `select()` instead.
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_warnings()` to see where this warning was generated.

#> # A tibble: 100 x 4
#>    index .actual .fitted .resid
#>    <int>   <dbl>   <dbl>  <dbl>
#>  1     1      88    88    0    
#>  2     2      84    85.1 -1.11 
#>  3     3      85    83.6  1.39 
#>  4     4      85    85.3 -0.277
#>  5     5      84    84.3 -0.277
#>  6     6      85    84.4  0.554
#>  7     7      83    83.8 -0.831
#>  8     8      85    83.9  1.11 
#>  9     9      88    87.7  0.277
#> 10    10      89    89.6 -0.554
#> # … with 90 more rows